v1.5 Enhancements and Bug Fixes
* More changes to the header text. For child themes, if a custom header
is uploaded, the front page will display the site title and description.
* Change "smooth scroll" arrow on home page with featured image to
scroll to the next section, whether that is a Page Top widget or the
page content itself.
* Site index page template (and 404 page) now supports Jetpack
Portfolios and Testimonials. Displays links in the Site Content (pages)
section. Also lists portfolio types and tag cloud if there are any.