Version 1.7.1 Quick Fix for libxml Error
Quick fix to suppress an error message about libxml invalid tags when
processing the site footer. This would only display when WordPress was
in debug mode (like on a test site), but it was annoying anyway. If
anyone cares, the "error" is really a warning that PHP's libxml doesn't
yet understand HTML5 tags. It still processes them just fine, though.